build: switch to src/app-only; remove legacy app/; add chat endpoints
