--- app/.well-known/appspecific/com.chrome.devtools.json/route.ts
@@ -1,3 +1,0 @@
-export function GET() {
-  return new Response("{}", { status: 200, headers: { "Content-Type": "application/json" } })
-}
