--- app/api/stripe/checkout/route.ts
@@ -1,4 +1,0 @@
-export const runtime = 'nodejs';
-export async function POST() {
-  return new Response('Deprecated: use /api/checkout', { status: 410 });
-}
